+++ app/src/main/java/com/example/myapplication_1/Fragments/EditMyAddressFragment.java
@@ -5,6 +5,7 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageButton;
+import android.widget.TextView;
 
 import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentTransaction;
@@ -16,6 +17,7 @@ public class EditMyAddressFragment extends Fragment {
     public static final String TAG = "EditMyAddressFragment";
 
     ImageButton button;
+    TextView textView;
 
     public static EditMyAddressFragment getInstance(Bundle args) {
 
@@ -35,8 +37,21 @@ public class EditMyAddressFragment extends Fragment {
                 container, false);
 
         button = view.findViewById(R.id.rl_edit_my_address_button);
+        textView = view.findViewById(R.id.rl_edit_my_address_delete);
 
         button.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                Bundle _args = new Bundle();
+                Fragment addressFragment = AddressFragment.getInstance(_args);
+
+                FragmentTransaction fragmentTransaction = getFragmentManager().beginTransaction();
+                fragmentTransaction.replace(R.id.ll_main, addressFragment);
+                fragmentTransaction.commit();
+            }
+        });
+
+        textView.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
                 Bundle _args = new Bundle();
