+++ app/src/main/res/layout/about_app.xml
@@ -109,6 +109,7 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_weight="1"
+        android:foreground="?android:selectableItemBackground"
         android:orientation="horizontal"
         android:background="@drawable/grey_line">
