+++ app/src/main/res/layout/cell_addresses.xml
@@ -9,6 +9,7 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:paddingLeft="20dp"
+        android:foreground="?android:selectableItemBackground"
         android:orientation="horizontal">
 
     <ImageButton
