+++ app/src/main/res/layout/cell_menu_1_1.xml
@@ -10,9 +10,8 @@
             android:layout_height="wrap_content"
             android:paddingTop="20dp"
             android:paddingBottom="20dp"
-            android:foreground="@android:color/transparent"
+            android:foreground="?android:selectableItemBackground"
             android:paddingLeft="20dp"
-            android:foregroundTint="@color/my_black"
             android:clickable="true"
             android:focusable="true"
             android:text="+7 963 377 88 44"
