+++ app/src/main/res/layout/cell_menu_1_6_2.xml
@@ -9,6 +9,7 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:paddingTop="15dp"
+        android:foreground="?android:selectableItemBackground"
         android:paddingBottom="15dp"
         android:background="@drawable/grey_line">
