+++ app/src/main/res/layout/cell_partners.xml
@@ -8,6 +8,7 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:paddingLeft="20dp"
+        android:foreground="?android:selectableItemBackground"
         android:orientation="horizontal">
 
         <TextView
