+++ app/src/main/res/layout/edit_my_address_fragment.xml
@@ -30,7 +30,7 @@
             android:src="@drawable/icon_strelka" />
 
         <TextView
-            android:id="@+id/rl_edit_my_address.delete"
+            android:id="@+id/rl_edit_my_address_delete"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_alignParentRight="true"
