+++ app/src/main/res/layout/payment_pick.xml
@@ -30,6 +30,7 @@
             android:id="@+id/rl_payment_pick_rl_visa"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
+            android:foreground="?android:selectableItemBackground"
             android:paddingLeft="15dp"
             android:paddingRight="15dp">
 
@@ -70,6 +71,7 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:orientation="horizontal"
+            android:foreground="?android:selectableItemBackground"
             android:paddingLeft="15dp"
             android:paddingRight="15dp">
 
@@ -108,6 +110,7 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:orientation="horizontal"
+            android:foreground="?android:selectableItemBackground"
             android:paddingLeft="15dp"
             android:paddingRight="15dp">
