+++ app/src/main/res/layout/settings.xml
@@ -46,6 +46,12 @@
                 android:layout_height="1dp"
                 android:background="#EDEDED"/>
 
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:orientation="vertical"
+        android:foreground="?android:selectableItemBackground">
+
                 <TextView
                     android:id="@+id/ll_settings_title1"
                     android:layout_width="match_parent"
@@ -67,5 +73,6 @@
                     android:paddingBottom="15dp"
                     android:background="@drawable/grey_line"
                     android:textColor="#C6C6C6"/>
+    </LinearLayout>
 
 </LinearLayout>
