+++ app/src/main/res/layout/tariffs.xml
@@ -8,7 +8,7 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_weight="0"
-        android:paddingLeft="15dp">
+        android:paddingLeft="10dp">
 
         <ImageButton
             android:id="@+id/rl_tariffs_button"
